// apps/api/src/child/enums/child.enums.ts
export enum Gender {
  Boy = 'Boy',
  Girl = 'Girl',
}

export enum Grade {
  Grade1 = 'Grade 1',
  Grade2 = 'Grade 2',
  Grade3 = 'Grade 3',
  Grade4 = 'Grade 4',
  Grade5 = 'Grade 5',
  Grade6 = 'Grade 6',
}

export enum Curriculum {
  IB = 'IB',
  Cambridge = 'Cambridge',
  Merdeka = 'Merdeka',
  Nasional = 'Nasional',
}

export enum PreferredLanguage {
  English = 'English',
  BahasaIndonesia = 'Bahasa Indonesia',
}